Compact template placement so any buildable airport fits
A template is placed as a single command, because placement preflights the
whole layout before executing and so cannot be split. That put a ceiling on
how many tiles a template could hold, and the ceiling sat below the largest
airport a player can build: station_spread tops out at 64, so a solid
airport reaches 64x64 = 4096 tiles.

Shrink the per-tile wire encoding from 8 bytes to 5. Tile offsets become one
byte each -- they are offsets inside a bounding box that station_spread
already limits -- and the five flag fields are bit-packed into two bytes.
A full-size layout is now 20494 bytes against a 32703-byte payload budget,
so MAX_TEMPLATE_TILES can be the structural maximum and no buildable
airport is ever too large to save.

Packing narrower-than-a-byte fields means a value too wide for its field
would survive a local execution and be truncated by a round trip over the
network, giving client and server different airports. Reject such tiles in
the command instead, so both paths always execute the same layout.

+++ src/modular_airport_gui.cpp
@@ -1016,7 +1016,7 @@ private:
 			data.height = static_cast<uint16_t>(GetModularCompoundPieceSize(piece.apt_gfx).height);
 			data.rotation = 0;
 			for (const ModularCompoundPieceTile &ct : compound) {
-				data.tiles.push_back({static_cast<uint16_t>(ct.dx), static_cast<uint16_t>(ct.dy), ct.gfx, 0, 0, false, 0x0F, 0});
+				data.tiles.push_back({static_cast<uint8_t>(ct.dx), static_cast<uint8_t>(ct.dy), ct.gfx, 0, 0, false, 0x0F, 0});
 			}
 
 			auto proc = [=](bool test, StationID to_join) -> bool {
